refactor: rename files to address-book
