+++ src/models/address-book/address-book.stub.js
