+++ src/models/address-book/index.js
@@ -1,0 +1,2 @@
+export * from './address-book.model';
+export * from './address-book.mock';
