--- src/models/address/index.js
@@ -1,2 +1,0 @@
-export * from './address.model';
-export * from './address.mock';
