+++ src/routes/v1/resources/index.js
@@ -9,4 +9,4 @@ export * from './governance';
 export * from './filestore';
 export * from './offer';
 export * from './statistics';
-export * from './address';
+export * from './address-book';
